refactor: Перенесла типы ActionItem в папку interface
